traduce las coordenadas mal

+++ practica2.py
@@ -244,6 +244,7 @@ def rotacion():
     #funcion que toma los puntos del array de coordenadas y los rota en funcion de la entrada de rotacion.
     #se rotan los elementos en funcion del centro del canvas
     #se debe de verificar que el array de coordenadas no este vacio
+
     if len(coordenadas) <= 0:
         return -1
     
@@ -257,11 +258,13 @@ def rotacion():
         x = coordenadas[i][0]
         y = coordenadas[i][1]
 
-        coordenadas_traducidas[i][0] = (x * math.cos(alfa_radianes)) - (y * math.sin(alfa_radianes))
-        coordenadas_traducidas[i][1] = (x * math.sin(alfa_radianes)) + (y * math.cos(alfa_radianes))
+        coordenadas_traducidas[i][0] = int( (x * math.cos(alfa_radianes)) - (y * math.sin(alfa_radianes)))
+        coordenadas_traducidas[i][1] = int( (x * math.sin(alfa_radianes)) + (y * math.cos(alfa_radianes)))
 
     for i in range(len(coordenadas)):
+        #TODO traduce mal las coordenadas
         coordenadas[i] = destraduccir_coordenadas(coordenadas_traducidas[i][0], coordenadas_traducidas[i][1])
+        
 
     #se limpia el canvas
     flush_canvas()
